move nested types to separate files
